Added Tangle and Weave Tools for Literate Programming
The oa-literate-weave will combine the projects components (weave the
pieces to create documentation0.
The oa-literate-tangle will extract the chunks from documentation and
update the executable program.
TODO: Explain each part in the documentation to generate the files,
rather than just getting the files and putting them in a master doc.
Use: When discussing the system with AI, by having the responses be able
to automatically be extracted and be put into place makes it quick to
prototype and workshop ideas. And using the master document, makes it
easier to reason, and allows for documentation to be added to explore
ideas. And allow the system to be built as ideas grow in bits and
pieces. Knuth says it makes us better programmers. I am starting to see
why :)