Added Diagram to show worktree-workspace morphism
The configuration makes branch versions isomorphic
not to the branch object, but what it represents.

A coalgebraic model allows for coherence
behind the unorthodox use of git branches,
and worktrees, as a composable component
of the system, as part of the ongoing exploration
into beating entropy with universal abstract structures
that allows the initial morphism to be repeated to
rebuild a system in chaos.